backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
psycopg2-binary
pydantic
python-multipart
aiofiles
python-dotenv
Pillow
PyPDF2
pdf2image
